backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pc_agent_backend"
version = "0.0.1"
description = "Local backend for the pc_agent personal browser memory + action agent."
requires-python = ">=3.11"
authors = [{ name = "adityaramkumar", email = "ramkumar.aditya@gmail.com" }]
license = { text = "MIT" }
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "pydantic>=2.9",
  "pydantic-settings>=2.6",
  "python-dotenv>=1.0",
  "google-genai>=0.8",
  "sqlite-vec>=0.1.6",
  "httpx>=0.27",
]
# NOTE on sqlite extension loading: sqlite-vec is loaded as a sqlite extension,
# so the running Python must have `Connection.enable_load_extension`. Linux
# stdlib Python always does. macOS Python.org builds do *not*, so use Homebrew
# Python (`brew install python@3.12`) or run `pip install pysqlite3-binary`
# manually if a wheel is available for your platform.

[project.optional-dependencies]
dev = [
  "ruff>=0.7",
  "mypy>=1.13",
  "pytest>=8.3",
  "pytest-asyncio>=0.24",
  "types-requests",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = [".venv", "dist", "build"]

[tool.ruff.lint]
select = [
  "E", "F", "W",  # pycodestyle / pyflakes
  "I",            # isort
  "B",            # bugbear
  "UP",           # pyupgrade
  "SIM",          # simplify
  "RUF",          # ruff-specific
]
ignore = [
  "E501",  # line length is enforced by formatter
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unreachable = true
disallow_untyped_decorators = false
plugins = []

[[tool.mypy.overrides]]
module = ["sqlite_vec.*", "google.genai.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra"
